:recycle: Update: Structure
