+++ model/__init__.py
@@ -1,7 +1,10 @@
 from .model import *
-from .train import *
 from .load_data import *
+from .train import *
 from .test import *
 
 
-__all__ = ['init_model', 'prepare_training', 'train', 'load_data', 'draw_gt', 'get_results', 'draw_res', 'test']
+__all__ = ['init_model',
+           'load_data',
+           'prepare_training', 'train',
+           'draw_gt', 'get_results', 'draw_res', 'test']
