+++ model/load_data.py
@@ -1,8 +1,68 @@
+import os
+from PIL import Image
+
+import numpy as np
+
 import torch
 
 from utils import utils
-from utils import get_transform, CustomizedDataset
+from utils import transforms as T
 
+
+def get_transform(train):
+    transforms = []
+    transforms.append(T.PILToTensor())
+    if train:
+        transforms.append(T.RandomHorizontalFlip(0.5))
+    return T.Compose(transforms)
+
+class CustomizedDataset(torch.utils.data.Dataset):
+    def __init__(self, root, transforms=None):
+        self.root = root
+        self.transforms = transforms
+        self.imgs = list(sorted(os.listdir(os.path.join(root, "images"))))
+        self.masks = list(sorted(os.listdir(os.path.join(root, "masks"))))
+
+    def __getitem__(self, idx):
+        img_path = os.path.join(self.root, "images", self.imgs[idx])
+        mask_path = os.path.join(self.root, "masks", self.masks[idx])
+        img = Image.open(img_path).convert("RGB")
+        mask = Image.open(mask_path)
+        mask = np.array(mask)
+        obj_ids = np.unique(mask)
+        obj_ids = obj_ids[1:]
+        masks = mask == obj_ids[:, None, None]
+        num_objs = len(obj_ids)
+        boxes = []
+        for i in range(num_objs):
+            pos = np.where(masks[i])
+            xmin = np.min(pos[1])
+            xmax = np.max(pos[1])
+            ymin = np.min(pos[0])
+            ymax = np.max(pos[0])
+            boxes.append([xmin, ymin, xmax, ymax])
+        boxes = torch.as_tensor(boxes, dtype=torch.float32)
+        labels = torch.as_tensor(obj_ids, dtype=torch.int64)
+        masks = torch.as_tensor(masks, dtype=torch.uint8)
+        image_id = torch.tensor([idx])
+        area = (boxes[:, 3] - boxes[:, 1]) * (boxes[:, 2] - boxes[:, 0])
+        iscrowd = torch.zeros((num_objs,), dtype=torch.int64)
+
+        target = {}
+        target["boxes"] = boxes
+        target["labels"] = labels
+        target["masks"] = masks
+        target["image_id"] = image_id
+        target["area"] = area
+        target["iscrowd"] = iscrowd
+
+        if self.transforms is not None:
+            img, target = self.transforms(img, target)
+
+        return img, target
+
+    def __len__(self):
+        return len(self.imgs)
 
 def load_data(TrainingDir, TestDir, batch_size=8, num_workers=16):
     TrainingDataset = CustomizedDataset(TrainingDir, get_transform(train=True))
